+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- CUMULUS-1488 - Update java client to allow use of CUMULUS_MESSAGE_ADAPTER_DIR environment variable/support layers
+
 ## [v1.2.9] - 2019-09-16
 ### Added
 - Updated CMA client to handle parameterized configuration, set execution env variable regardless of message format
-